Update handling of repo files for CentOS 8.3
In the centos element, the option DIB_DISTRIBUTION_MIRROR can be
used to set a local mirror for package repos.  The implementation
of this was broken by CentOS 8.3 renaming package repo files.

This fix updates to the new package file names, now provided by the
centos-linux-repos RPM, and assumes these will apply going forwards
from 8.3 on the CentOS 8 release cycle.

Change-Id: I985f7863def5247f00761e8509082affdab262cf